Time: 0 ms (100%), Space: 71.6 MB (80.79%) - LeetHub

+++ 0053-maximum-subarray/0053-maximum-subarray.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int maxSubArray(vector<int>& nums) {
-        long sum=0;
+        long long sum=0;
         long maxi=INT_MIN;
         int n=nums.size();
         for(int i=0;i<n;i++){
